+++ CHANGELOG.md
@@ -15,6 +15,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - iBoM:
   - `forced_name` option to force the name displayed at the top left corner
     (#470)
+- Blender export:
+  - Support for pcb2blender v2.6 (Blender 3.5.1)
 
 ### Fixed
 - Rotated polygons and text used in the worksheet (#466)